Simplify GSSAPI key exchange algorithm handling
It doesn't make much sense to have "ssh -Q kex" output names which are
only prefixes and which thus can't be used in the KexAlgorithms option.
If this is needed then we should invent a new -Q option name for it
instead (perhaps "gssapi-kex"), but that can probably wait.

Keeping track of algorithm prefixes in a separate array also simplifies
kex_alg_by_name.