Move remote_hostname to canohost.[ch]
Fedora previously moved this to openbsd-compat/port-linux.[ch] for use
by the implementation of the GSSAPITrustDns option, but that's incorrect
because the function isn't specific to Linux and is needed when building
on other kernels too.  canohost.[ch] seems reasonable (indeed, that's
where it used to live upstream), so move it there instead.