Add GSSAPICleanupCreds as a compatibility alias
The option is called GSSAPICleanupCredentials, but Debian has had a
"GSSAPICleanupCreds" compatibility alias for a long time, I think
because some very old versions of the GSSAPI key exchange patch used
that spelling instead and it was easier to add two lines to the patch
than to reliably migrate configurations.  It seems harmless enough.